fix: move agent-task status evidence support to core
